Initial commit of MyService to run any app as a service.
MyService uses Apache ProcRun to enable any application to be run as a service. It is preconfigured for running a VirtualBox VM as a service and saving its state on shutdown. The start and stop commands are in the myservice.properties file. The original idea came from Christopher Pierce's blog entry "Java as Windows Service with Apache Commons Daemon", http://blog.platinumsolutions.com/node/234